refactor: move files to correct folder. Use `framework` folder for framework-specific code, use `components/user` form user-specific API business logic
