+++ jest.config.ts
@@ -3,6 +3,7 @@ export default {
     clearMocks: true,
     coverageProvider: 'v8',
     preset: 'ts-jest',
+    verbose: true,
     testMatch: [
         '**/__tests__/**/*.test.ts',
     ],
