+++ src/framework/error/HttpBadRequestError.ts
